% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enshuman.R
\docType{package}
\name{enshuman}
\alias{enshuman}
\title{enshuman}
\description{
Human gene annotation datasets to use with the 'topr' package
}
\section{datasets}{

The two datasets are:
\itemize{
\item \code{\link{hg38}} Gene information from genome build GRCh38.p14
\item \code{\link{hg37}} Gene information from genome build GRCh37.p13
}
}

\examples{
library(enshuman)
head(hg38)
head(hg37)
}
