% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hg37}
\alias{hg37}
\title{Ensembl genes build GRCh37.87}
\format{
A data frame  with 55,882 rows and 7 variables:
\describe{
\item{chrom}{chromosome}
\item{gene_start}{genetic position of gene start}
\item{gene_end}{genetic position of gene end}
\item{gene_symbol}{The name of the gene}
\item{biotype}{the biotype of the gene}
\item{exon_chromstart}{genetic positions of exon start}
\item{exon_chromend}{genetic position of exon end}
}
}
\usage{
hg37
}
\description{
wget https://ftp.ensembl.org/pub/grch37/current/gtf/homo_sapiens/Homo_sapiens.GRCh37.87.gtf.gz
}
\keyword{datasets}
