% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hg38}
\alias{hg38}
\title{Ensembl genes build GRCh38.111}
\format{
A data frame  with 41,016 rows and 7 variables:
\describe{
\item{chrom}{chromosome}
\item{gene_start}{genetic position of gene start}
\item{gene_end}{genetic position of gene end}
\item{gene_symbol}{the name of the gene}
\item{biotype}{the biotype of the gene}
\item{exon_chromstart}{genetic positions of exon start}
\item{exon_chromend}{genetic position of exon end}
}
}
\usage{
hg38
}
\description{
wget https://ftp.ensembl.org/pub/current/gtf/homo_sapiens/Homo_sapiens.GRCh38.111.gtf.gz
}
\keyword{datasets}
